// FrontEnd/src/app/models/info-device/info-device.ts
export class InfoDevice {
  Device: string = "N/D";
  DriverData: string = "N/D";
  EndDate: string = "N/D";
  Fuel: string = "0";
  Image: string = "N/D";
  ODO: string = "N/D";
  Score: string = "N/D";
  StartDate: string = "N/D";
  Time: string = "N/D";
  estatus: number = 0;
  VehicleName: string = "N/D";
  Modelo: string = "N/D";
  Marca: string = "N/D";
  Version: string = "N/D";
  TipoVehiculo: string = "N/D";
  VIN: string = "N/D";
  Poliza: string = "N/D";
  Placas: string = "N/D";
  Batery: string = "N/D";
  Label: string = "N/D";
}